requirements.txt
numpy==1.25.2
Pillow==10.0.0
pydantic_core==2.6.1
pytest==7.4.1